refactor: carve graph out to single file
