+++ entropix/local/tokenizer.py
@@ -1,6 +1,7 @@
 import os
 from logging import getLogger
 from pathlib import Path
+import urllib
 from typing import (
   AbstractSet,
   cast,
